Changed configuration of database files. Issue nb 5.
