+++ app/config.py
@@ -1,0 +1,7 @@
+
+database=(
+    DATABASE='',
+    SECRET_KEY='',
+    USERNAME='',
+    PASSWORD=''
+)
