+++ app/script.py
@@ -1,64 +1,25 @@
 from flask import Flask, render_template, request, session, g, redirect, url_for, abort, flash
 import sqlite3
 import os
+from config import database
+from database import *
 
 app=Flask(__name__)
 app.config.from_object(__name__)
 
-app.config.update(dict(
-    DATABASE=os.path.join(app.root_path, 'app.db'),
-    SECRET_KEY='development key',
-    USERNAME='admin',
-    PASSWORD='default'
-))
+app.config.update(dict(database))
+
 app.config.from_envvar('APP_SETTINGS', silent=True)
-
-
-def connect_db():
-    """Connects to the specific database."""
-    rv = sqlite3.connect(app.config['DATABASE'])
-    rv.row_factory = sqlite3.Row
-    return rv
-
-def init_db():
-    db = get_db()
-    with app.open_resource('database.sql', mode='r') as f:
-        db.cursor().executescript(f.read())
-    db.commit()
-
-@app.cli.command('initdb')
-def initdb_command():
-    """Initializes the database."""
-    init_db()
-    print('Initialized the database.')
-
-
-
-def get_db():
-    """Opens a new database connection if there is none yet for the
-    current application context.
-    """
-    if not hasattr(g, 'sqlite_db'):
-        g.sqlite_db = connect_db()
-    return g.sqlite_db
-
-
-@app.teardown_appcontext
-def close_db(error):
-    """Closes the database again at the end of the request."""
-    if hasattr(g, 'sqlite_db'):
-        g.sqlite_db.close()
-
-
 
 
 @app.route('/')
 def home():
     db = get_db()
-    cur = db.execute('select Name, Type from Event order by Id desc')
+    cur = db.execute('select Name, Type from Event limit 5 order by Id desc')
+    cur2=db.execute('select distinct City from Localization order by City')
     events = cur.fetchall()
-    return render_template('main.html', events=events)
-
+    cities=cur2.fetchall()
+    return render_template('main.html', events=events, cities=cities)
 
 
 if __name__ == "__main__":
